+++ main-camp/src/app/boards/new/page.tsx
@@ -1,8 +1,16 @@
+"use client"
+
+import { useIsEdit } from "@/commons/isEditProvider"
 import BoardsWrite from "@/components/boards-write"
+import { useEffect } from "react"
 
 const BoardNewPage = () => {
+    const { setIsEdit } = useIsEdit()
+    useEffect(()=>{
+        setIsEdit(false)
+    },[])
     return (
-        <BoardsWrite isEdit={false}/>
+        <BoardsWrite />
     )
 }
