+++ main-camp/src/components/boards-detail/style.module.css
@@ -1,8 +1,8 @@
-/* /boards/new */
+/* /boards/detail */
 
 /* main */
 
-.new_main {
+.detail_main {
     max-width: 1280px;
     min-width: 781px;
     margin-left: 320px;
@@ -11,49 +11,95 @@
     /* height: 100vh; */
 }
 
-.board_container {
-    width: 1280;
-    gap: 40px;
-    opacity: 1;
-}
+/* header */
 
-.write_form_container {
-    display: flex;
-    flex-direction: column;
-
-    width: 100%;
-    height: 1352px;
-    gap: 40px;
-    opacity: 1;
-}
-
-.write_confirm_container {
-    width: 100%;
-    height: 48px;
+.header_1280w_80h {
+    width: 1280px;
+    height: 80px;
     gap: 16px;
     opacity: 1;
 }
 
-/* write_form */
-
-.write_form_80h {
-    width: 100%;
-    height: 80px;
-    gap: 40px;
+.header_top {
+    width: 1280px;
+    height: 24px;
+    gap: 54px;
     opacity: 1;
 }
 
-.write_form_192h {
-    width: 100%;
-    height: 192px;
+.header_bottom {
+    width: 56px;
+    height: 24px;
     gap: 8px;
     opacity: 1;
 }
 
-.write_form_368h {
+.detail_profile {
     width: 100%;
-    height: 368px;
-    gap: 8px;
+    height: 24px;
+    gap: 4px;
     opacity: 1;
 }
 
+/* detail content */
+
+.profile_img {
+    width: 24px;
+    height: 24px;
+    opacity: 1;
+    background: var(--gray-100, rgba(228, 228, 228, 1));
+    border-radius: 100px;
+}
+
+.img_24w_24h {
+    width: 24px;
+    height: 24px;
+    opacity: 1;
+}
+
+.detail_video {
+    width: 100%;
+    padding-top: 24px;
+    padding-bottom: 24px;
+    gap: 10px;
+    opacity: 1;
+    background: var(--gray-50, rgba(242, 242, 242, 1));
+}
+
+.bad_good_48h {
+    width: 100%;
+    height: 48px;
+    gap: 24px;
+    opacity: 1;
+}
+
+.bad_good_btn_frame {
+    width: 24px;
+    height: 48px;
+    gap: 4px;
+    opacity: 1;
+    cursor: pointer; 
+}
+
+.detail_update {
+    width: 100%;
+    height: 40px;
+    gap: 24px;
+    opacity: 1;
+}
+
+.detail_update > button {
+    width: 105px;
+    height: 40px;
+    padding-top: 8px;
+    padding-right: 12px;
+    padding-bottom: 8px;
+    padding-left: 12px;
+    gap: 8px;
+    opacity: 1;
+    border-radius: 8px;
+    border-width: 1px;
+    background: var(--gray-w, rgba(255, 255, 255, 1));
+    border: 1px solid var(--gray-B, rgba(0, 0, 0, 1));
+    cursor: pointer;
+}
