+++ main-camp/src/components/boards-list/type.ts
@@ -1,0 +1,13 @@
+export interface IFetchBoardsData {
+    fetchBoards: {
+        _id: string,
+        title: string,
+        writer?: string,
+        contents: string,
+        createdAt: string
+    }[]
+}
+
+// interface IFetchBoardsCount {
+//     fetchBoardsCount: number
+// }
