+++ main-camp/src/components/boards-write/writeInput/type.ts
@@ -1,0 +1,17 @@
+import { ChangeEventHandler } from "react";
+
+type fetchBoard = {
+    fetchBoard : {
+        writer: string,
+        title: string,
+        contents: string
+    }
+}
+
+export interface IWriteInputProps {
+    data?: fetchBoard,
+    label: string,
+    placeholder?: string,
+    errMsg?: string
+    setState?: ChangeEventHandler<HTMLInputElement | HTMLTextAreaElement> | undefined,
+}
