+++ main-camp/tsconfig.json
@@ -21,6 +21,6 @@
       "@/*": ["./src/*"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "src/components/boards/writeButton", "src/components/boards/writeInput"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "src/components/boards-write/writeButton", "src/components/boards-write/writeInput"],
   "exclude": ["node_modules"]
 }
